# Add faster `df_cast_loop()` path to `df_cast_opts()`