Closes #1070 

Adds a fast path to `df_cast_opts()` when `x` and `to` have the same names. Very similar to `df_ptype2()`. Speed up is ~2x when the fast path is used, but I could see this being a very common case.

---


These examples take 10000 data frames of different types and cast them to the same common type.

``` r
library(vctrs)

vec_cast_common_lst <- function(x) {
  vec_cast_common(!!!x)
}
```

These all share the same `names()` SEXP, and the fast path is used.

```r
x <- rep_len(list(data.frame(x = 1, y = 2)), 10000)

# master
bench::mark(vec_cast_common_lst(x), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(x)   15.5ms   19.9ms      49.5     160KB     33.0

# this PR
bench::mark(vec_cast_common_lst(x), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(x)   7.84ms    9.3ms      105.     160KB     21.5
```

These all have different `names()` SEXP, and the fast path is used.

```r
y <- replicate(10000, data.frame(x = 1, y = 2), simplify = FALSE)

# master
bench::mark(vec_cast_common_lst(y), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(y)   22.2ms   27.7ms      35.5     156KB     34.1

# this PR
bench::mark(vec_cast_common_lst(y), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(y)   14.8ms   18.3ms      55.6     156KB     16.6
```

Introduction of `z` forces a common type of `df<x, y, z, w>`, so now none of the data frames will use the fast path. This is mainly to show that there is minimal overhead in this approach when it isn't used.

```r
z <- data.frame(z = 1, w = 2)
z <- c(list(z), x)

# master
bench::mark(vec_cast_common_lst(z), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(z)   29.4ms   51.6ms      20.5     156KB     55.0

# this PR
bench::mark(vec_cast_common_lst(z), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast_common_lst(z)     35ms   52.8ms      20.6     156KB     178.
```